anime_upscaler_v2/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Image processing and AI
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# System dependencies (install separately)
# ffmpeg - video processing
# waifu2x-ncnn-vulkan - anime upscaling (optional)
# realesrgan-ncnn-vulkan - high-quality upscaling (optional)

# Note: AI models require separate installation:
# - Waifu2x: https://github.com/nihui/waifu2x-ncnn-vulkan
# - Real-ESRGAN: https://github.com/xinntao/Real-ESRGAN
# - FFmpeg: https://ffmpeg.org/download.html